<Fix> Rename files, add more enum, delete outdated routes
